+++ src/configurations/index.ts
@@ -2,6 +2,7 @@ export type { CausaConfiguration } from './causa.js';
 export type { DockerConfiguration } from './docker.js';
 export type { EventsConfiguration } from './events.js';
 export type { InfrastructureConfiguration } from './infrastructure-project.js';
+export type { ModelConfiguration } from './model.js';
 export type { OpenApiConfiguration } from './openapi.js';
 export type { ServerlessFunctionsConfiguration } from './serverless-functions-project.js';
 export type {