+++ src/functions/event-topic/list-referenced-in-project-service-container.spec.ts
@@ -1,17 +1,57 @@
 import { NoImplementationFoundError } from '@causa/workspace/function-registry';
 import { createContext } from '@causa/workspace/testing';
 import 'jest-extended';
-import { EventTopicListReferencedInProject } from '../../definitions/index.js';
+import {
+  EventTopicList,
+  EventTopicListReferencedInProject,
+  MissingEventTopicDefinitionsError,
+  type EventTopicDefinition,
+} from '../../definitions/index.js';
 import { EventTopicListReferencedInProjectForServiceContainer } from './list-referenced-in-project-service-container.js';
 
 describe('EventTopicListReferencedInProjectForServiceContainer', () => {
+  const myFirstEventDefinition = {
+    id: 'my.first-event.v1',
+    formatParts: { domain: 'my', name: 'first-event', version: 'v1' },
+    schemaFilePath: 'my/first-event/v1.yaml',
+  };
+  const mySecondEventDefinition = {
+    id: 'my.second-event.v1',
+    formatParts: { domain: 'my', name: 'second-event', version: 'v1' },
+    schemaFilePath: 'my/second-event/v1.yaml',
+  };
+  const myOtherEventDefinition = {
+    id: 'my.other-event.v1',
+    formatParts: { domain: 'my', name: 'other-event', version: 'v1' },
+    schemaFilePath: 'my/other-event/v1.yaml',
+  };
+
+  class EventTopicListMock extends EventTopicList {
+    async _call(): Promise<EventTopicDefinition[]> {
+      return [
+        myFirstEventDefinition,
+        mySecondEventDefinition,
+        myOtherEventDefinition,
+      ];
+    }
+
+    _supports(): boolean {
+      return true;
+    }
+  }
+
+  const baseConfiguration = {
+    workspace: { name: '🏷️' },
+    project: { type: 'serviceContainer', language: 'ts', name: '🧪' },
+  };
+
   it('should return empty lists', async () => {
     const { context } = createContext({
-      configuration: {
-        workspace: { name: '🏷️' },
-        project: { type: 'serviceContainer', language: 'ts', name: '🧪' },
-      },
-      functions: [EventTopicListReferencedInProjectForServiceContainer],
+      configuration: baseConfiguration,
+      functions: [
+        EventTopicListReferencedInProjectForServiceContainer,
+        EventTopicListMock,
+      ],
     });
 
     const actualTopics = await context.call(
@@ -25,26 +65,13 @@ describe('EventTopicListReferencedInProjectForServiceContainer', () => {
   it('should return the consumed and produced topic', async () => {
     const { context } = createContext({
       configuration: {
-        workspace: { name: '🏷️' },
-        project: { type: 'serviceContainer', language: 'ts', name: '🧪' },
+        ...baseConfiguration,
         serviceContainer: {
           triggers: {
-            myFirstTrigger: {
-              type: 'event',
-              topic: 'my.first-event.v1',
-            },
-            mySecondTrigger: {
-              type: 'event',
-              topic: 'my.second-event.v1',
-            },
-            myThirdTrigger: {
-              type: 'event',
-              topic: 'my.second-event.v1',
-            },
-            notAnEventTrigger: {
-              type: 'cron',
-              schedule: '* * * * *',
-            },
+            myFirstTrigger: { type: 'event', topic: 'my.first-event.v1' },
+            mySecondTrigger: { type: 'event', topic: 'my.second-event.v1' },
+            myThirdTrigger: { type: 'event', topic: 'my.second-event.v1' },
+            notAnEventTrigger: { type: 'cron', schedule: '* * * * *' },
           },
           outputs: {
             eventTopics: [
@@ -55,7 +82,10 @@ describe('EventTopicListReferencedInProjectForServiceContainer', () => {
           },
         },
       },
-      functions: [EventTopicListReferencedInProjectForServiceContainer],
+      functions: [
+        EventTopicListReferencedInProjectForServiceContainer,
+        EventTopicListMock,
+      ],
     });
 
     const actualTopics = await context.call(
@@ -64,21 +94,44 @@ describe('EventTopicListReferencedInProjectForServiceContainer', () => {
     );
 
     expect(actualTopics).toEqual({
-      consumed: expect.toContainAllValues([
-        'my.first-event.v1',
-        'my.second-event.v1',
+      consumed: expect.toIncludeSameMembers([
+        myFirstEventDefinition,
+        mySecondEventDefinition,
       ]),
-      produced: expect.toContainAllValues([
-        'my.first-event.v1',
-        'my.other-event.v1',
+      produced: expect.toIncludeSameMembers([
+        myFirstEventDefinition,
+        myOtherEventDefinition,
       ]),
     });
+  });
+
+  it('should throw if a non-existing event topic is referenced', async () => {
+    const { context } = createContext({
+      configuration: {
+        ...baseConfiguration,
+        serviceContainer: {
+          triggers: {
+            myTrigger: { type: 'event', topic: 'my.non-existing-event.v1' },
+          },
+        },
+      },
+      functions: [
+        EventTopicListReferencedInProjectForServiceContainer,
+        EventTopicListMock,
+      ],
+    });
+
+    const actualPromise = context.call(EventTopicListReferencedInProject, {});
+
+    await expect(actualPromise).rejects.toThrow(
+      MissingEventTopicDefinitionsError,
+    );
   });
 
   it('should not handle a project type other than service container', async () => {
     const { context } = createContext({
       configuration: {
-        workspace: { name: '🏷️' },
+        ...baseConfiguration,
         project: { type: 'function', language: 'ts', name: '🧪' },
       },
       functions: [EventTopicListReferencedInProjectForServiceContainer],